+++ demo/segmentation/index.js
@@ -10,8 +10,6 @@
 import * as H from '../../dist/human.esm.js'; // equivalent of @vladmandic/Human
 
 const humanConfig = { // user configuration for human, used to fine-tune behavior
-  // backend: 'wasm',
-  // wasmPath: 'https://cdn.jsdelivr.net/npm/@tensorflow/tfjs-backend-wasm@3.20.0/dist/',
   modelBasePath: 'https://vladmandic.github.io/human-models/models/',
   filter: { enabled: true, equalization: false, flip: false },
   face: { enabled: false },