+++ test/test-backend-node-wasm.js
@@ -3,7 +3,7 @@ const tf = require('@tensorflow/tfjs'); // wasm backend requires tfjs to be load
 const wasm = require('@tensorflow/tfjs-backend-wasm'); // wasm backend does not get auto-loaded in nodejs
 const { Canvas, Image } = require('canvas'); // eslint-disable-line node/no-extraneous-require, node/no-missing-require
 const H = require('../dist/human.node-wasm.js');
-const test = require('./test-node-main.js').test;
+const test = require('./test-node-main.js');
 
 H.env.Canvas = Canvas; // requires monkey-patch as wasm does not have tf.browser namespace
 H.env.Image = Image; // requires monkey-patch as wasm does not have tf.browser namespace
@@ -35,12 +35,16 @@ const config = {
 };
 
 async function main() {
-  wasm.setWasmPaths(config.wasmPath);
-  await tf.setBackend('wasm');
+  wasm.setWasmPaths(config.wasmPath, true);
+  const ok = await tf.setBackend('wasm');
+  if (!ok) {
+    test.log('error', 'failed: setwasmpath', config.wasmPath);
+    return;
+  }
   await tf.ready();
   H.env.updateBackend();
   log.info(H.env.wasm, config.wasmPath);
-  test(H.Human, config);
+  test.test(H.Human, config);
 }
 
 if (require.main === module) main();