+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # @vladmandic/human  
 
-  Version: **2.10.3**  
+  Version: **2.11.0**  
   Description: **Human: AI-powered 3D Face Detection & Rotation Tracking, Face Description & Recognition, Body Pose Tracking, 3D Hand & Finger Tracking, Iris Analysis, Age & Gender & Emotion Prediction, Gesture Recognition**  
   
   Author: **Vladimir Mandic <mandic00@live.com>**  
@@ -9,7 +9,7 @@
   
 ## Changelog
   
-### **HEAD -> main** 2022/09/25 mandic00@live.com
+### **HEAD -> main** 2022/09/27 mandic00@live.com
 
 - create funding.yml
 - fix rotation interpolation