+++ CHANGELOG.md
@@ -9,7 +9,7 @@
   
 ## Changelog
   
-### **HEAD -> main** 2022/09/27 mandic00@live.com
+### **HEAD -> main** 2022/09/29 mandic00@live.com
 
 - create funding.yml
 - fix rotation interpolation