+++ CHANGELOG.md
@@ -1,6 +1,6 @@
 # @vladmandic/human  
 
-  Version: **2.11.0**  
+  Version: **2.11.1**  
   Description: **Human: AI-powered 3D Face Detection & Rotation Tracking, Face Description & Recognition, Body Pose Tracking, 3D Hand & Finger Tracking, Iris Analysis, Age & Gender & Emotion Prediction, Gesture Recognition**  
   
   Author: **Vladimir Mandic <mandic00@live.com>**  
@@ -9,8 +9,12 @@
   
 ## Changelog
   
-### **HEAD -> main** 2022/09/29 mandic00@live.com
+### **HEAD -> main** 2022/10/09 mandic00@live.com
 
+
+### **origin/main** 2022/10/02 mandic00@live.com
+
+- add human.webcam methods
 - create funding.yml
 - fix rotation interpolation